requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
qdrant-client==0.11.9
openpyxl==3.1.2
python-dotenv==1.0.0
aiosqlite==0.19.0
bcrypt==4.0.1
cryptography==41.0.5
pydantic>=1.8,<2.0
starlette==0.27.0
typing-extensions>=4.8.0
openai==1.12.0 